Refactor contextual menu and add delete feature for recipe
